+++ src/views/Tag/Tags.vue
@@ -42,17 +42,8 @@
               <md-icon>minimize</md-icon>
             </div>
           </md-table-cell>
-          <md-table-cell md-label="Color"
-            ><md-chip
-              v-if="item.color"
-              :style="`background-color: ${
-                item.color
-              } !important; color: ${getFontColor(item.color)} !important`"
-              >{{ item.color }}</md-chip
-            >
-            <md-chip v-else :style="`background-color: #f9f9f9 !important;`"
-              >#f9f9f9</md-chip
-            >
+          <md-table-cell md-label="Color">
+            <TagBoxChip :color="item.color" :text="item.color"></TagBoxChip>
           </md-table-cell>
         </md-table-row>
       </md-table>
@@ -129,6 +120,7 @@ import firebase from "firebase";
 import { Tag } from "@/types/garnbarn/Tag";
 import { GetAllTagApiNextFunctionWrapper } from "@/types/GarnBarnApi/GarnBarnApiResponse";
 import UserProfileIcon from "@/components/UserProfileIcon.vue";
+import TagBoxChip from "@/components/Tag/TagBoxChip.vue";
 
 @Component({
   components: {
@@ -136,6 +128,7 @@ import UserProfileIcon from "@/components/UserProfileIcon.vue";
     DialogBoxComponent,
     Create,
     UserProfileIcon,
+    TagBoxChip,
   },
 })
 export default class Tags extends Vue {
@@ -162,22 +155,6 @@ export default class Tags extends Vue {
       this.getNextData = apiResponse.data.next;
       loadingDialogBox.dismiss();
     });
-  }
-
-  hexToRgb(hex: string): Array<number> {
-    const r = parseInt(hex.slice(1, 3), 16);
-    const g = parseInt(hex.slice(3, 5), 16);
-    const b = parseInt(hex.slice(5, 7), 16);
-    return [r, g, b];
-  }
-
-  getFontColor(hex: string): string {
-    const rgb = this.hexToRgb(hex);
-    if (rgb[0] * 0.299 + rgb[1] * 0.587 + rgb[2] * 0.114 > 186) {
-      return "#000000";
-    } else {
-      return "#ffffff";
-    }
   }
 
   processNext(): void {
